revamp fake user creation
